requirements.txt
Flask==3.1.2
gunicorn==23.0.0
google-auth==2.42.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.185.0
requests==2.32.3
setuptools
wheel

